+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Value objects were made serializable: `Year`, `Date`, `Time`, `LocalDateTime`, `ZonedDateTime`, `TimeZone`,
   `TimeZoneId`, `Offset`.
 
+### Changed
+- (BC Break) Dependency `litgroup/enumerable` was updated to `^0.6.0`.
+
 ## v0.1.1 - 2016-10-10
 ### Added
 - Added shortcut factory `ZonedDateTime::nowUtc()` which returns current time in the UTC timezone.